refactor(modal): replace implementation of closing modal by Esc key press
