+++ src/components/modal/modal.tsx
@@ -1,19 +1,37 @@
-import React, { JSX } from 'react';
+import React, { JSX, useEffect } from 'react';
 import s from './modal.module.css';
 import xmark from '../../images/icons/xmark.svg';
 import ModalOverlay from './modal-overlay/modal-overlay';
+import { isEscKeyPressed } from '../../utils/utils';
 
 type ModalProps = {
   title?: string;
   children: React.ReactNode;
   onClose: () => void;
+  isModalOpen: boolean;
 };
 
-function Modal({ title, children, onClose }: ModalProps): JSX.Element {
+function Modal({ title, children, onClose, isModalOpen }: ModalProps): JSX.Element {
   const handleCloseBtnClick = (e: React.MouseEvent) => {
     e.stopPropagation();
     onClose();
   };
+
+  useEffect(() => {
+    const handleEscKeyPress = (e: KeyboardEvent) => {
+      if (isEscKeyPressed(e)) {
+        onClose();
+      }
+    };
+
+    if (isModalOpen) {
+      window.addEventListener('keydown', handleEscKeyPress);
+    }
+
+    return () => {
+      window.removeEventListener('keydown', handleEscKeyPress);
+    };
+  }, [isModalOpen]);
 
   return (
     <div className={s.modal}>
